Fix examples, add missing sample formats, add h264+aac example

+++ examples/encoding-multiple.go
@@ -136,7 +136,7 @@ func main() {
 	}
 
 	var srcFrame *Frame
-	j := 0
+	j := int64(0)
 
 	for srcFrame = range GenSyntVideoNewFrame(320, 200, AV_PIX_FMT_YUV420P) {
 		srcFrame.SetPts(j)
